Fix problem with timer requests being in the past
In some situations (particularly when running a PanelDue or other UART
based display) requests to set a timer call back can be delayed such
that the requested time is now in the past. This means that the event
will not fire until the timer wraps around. This can cause heaters
(which use the timer API for software based PWM) to stay on, or appear
not to power on. The problem is that the call to set the timer can be
interrupted in some situations by a higher priority interrupt, or simply
the timer can tick before the new value is set.

This fix takes steps to ensure that the set timer target is always ahead
of the current timer value. In some cases this may result in an extra
timer interrupt (if the previous set value actually triggered an event)
and may also introduce a 1uS delay to the target time. But the current
code will in effect ignore the extra interrupt, and the delay is
probably will hopefully not cause problems.