Modified us ticker to use RIT Timer
These changes switch the uS timer code from using TIMER3 to using th RIT
timer. This frees up TIMER3 to allow the soft UART code to run at the
same time as the ADC filter.

Unfortunately the RIT timer does not have a pre-scaler option which
measn that we have to run the RIT timer at full PCLK rather than scaling
it to be 1 tick per uS. To preserve the external API we post scale the
tick values, however this means that the longest delta that can be set
is reduced form 2125 seconds to 85 seconds. This is not an issue with
the way that this code is used at present (for software PWM generation),
but should be noted if the code is used for other purposes.