Add performance stats and temporary fix
Have seen a few reports of watchdog events with the stack showing that
the code being executed was the Software PWM interrupt handler. I have
not been able to reproduce the problem, but adding code to ensure that
there is a minimum gap between interrupts seems to help those that have
reported the problem. This branch contains instrumented versions of the
new and old code.