Allow different interrupts for DMA completion
These changes allow different devices sharing the DMA controller
to generate completion interrupts that use different levels. This
allows the DMA device to operate at say a high priority but the
completion interrupt processing for the devices using it can operate
at a lower priority. In particular they can operate at a priority
that allows the use of RTOS functions.