Fix watchdog timeout when enabling PanelDue UART
When the UART was enabled with a PanelDue connected the firmware would
reset sometimes with a watchdog timeout. This seemes to have been caused
by a framing/parity/overrun error not being cleared by the normal read
interrupt handler and so the inetrrupt repeated over and over. The
current fix is to clear any such errors before enabling interrupts.