Hack to fix CDC full packet bug
The CDC specification indicates that transmissions to a host should be
terminated by either a short packet or a zero length packet. Some hosts
will not pass data up to the application layer until the end of data is
detected. The existing code takes no action to handle the sending of a
zlp in the case when the last packet is full length. Previously RRF
effectively sent only 1 character per packet (so all packets were
short), changes to mae this more efficient resulted in data being
displayed appear to hang. This hack works around the problem by using a
buffer size that ensures all packets are short (in this case 63 bytes or
less).